+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# 1.3.1
+readAsync works, but cancelAsync does not
+
+this means that if you want to use asynchronous operations they are unstoppable (I have no clue why it does not work)
+
 # 1.3.0
 Update packages, simple fixes (prepare for async)